refactor(bpmn-model): extract user task tests from process builder
